requirements.txt
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
connexion==2.14.2
Flask==2.2.5
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
MarkupSafe==2.1.5
packaging==24.0
python-dateutil==2.6.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
swagger-ui-bundle==0.0.9
urllib3==2.2.1
uuid==1.30
Werkzeug==2.2.3
wsproto==1.2.0
